Added `rf_proj_raster`
