+++ core/src/main/scala/org/locationtech/rasterframes/functions/package.scala
