+++ core/src/main/scala/org/locationtech/rasterframes/package.scala
