+++ core/src/test/scala/org/locationtech/rasterframes/RasterFunctionsSpec.scala
@@ -21,18 +21,11 @@
 
 package org.locationtech.rasterframes
 
-import java.io.ByteArrayInputStream
-
-import geotrellis.raster
 import geotrellis.raster._
-import geotrellis.raster.render.ColorRamps
 import geotrellis.raster.testkit.RasterMatchers
-import javax.imageio.ImageIO
 import org.apache.spark.sql.Encoders
 import org.apache.spark.sql.functions._
 import org.locationtech.rasterframes.expressions.accessors.ExtractTile
-import org.locationtech.rasterframes.model.TileDimensions
-import org.locationtech.rasterframes.stats._
 import org.locationtech.rasterframes.tiles.ProjectedRasterTile
 
 class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
@@ -41,74 +34,6 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
 
   implicit val pairEnc = Encoders.tuple(ProjectedRasterTile.prtEncoder, ProjectedRasterTile.prtEncoder)
   implicit val tripEnc = Encoders.tuple(ProjectedRasterTile.prtEncoder, ProjectedRasterTile.prtEncoder, ProjectedRasterTile.prtEncoder)
-
-  describe("constant tile generation operations") {
-    val dim = 2
-    val rows = 2
-
-    it("should create a ones tile") {
-      val df = (0 until rows).toDF("id")
-        .withColumn("const", rf_make_ones_tile(dim, dim, IntConstantNoDataCellType))
-      val result = df.select(rf_tile_sum($"const") as "ts").agg(sum("ts")).as[Double].first()
-      result should be (dim * dim * rows)
-    }
-
-    it("should create a zeros tile") {
-      val df = (0 until rows).toDF("id")
-        .withColumn("const", rf_make_zeros_tile(dim, dim, FloatConstantNoDataCellType))
-      val result = df.select(rf_tile_sum($"const") as "ts").agg(sum("ts")).as[Double].first()
-      result should be (0)
-    }
-
-    it("should create an arbitrary constant tile") {
-      val value = 4
-      val df = (0 until rows).toDF("id")
-        .withColumn("const", rf_make_constant_tile(value, dim, dim, ByteConstantNoDataCellType))
-      val result = df.select(rf_tile_sum($"const") as "ts").agg(sum("ts")).as[Double].first()
-      result should be (dim * dim * rows * value)
-    }
-  }
-
-  describe("cell type operations") {
-    it("should convert cell type") {
-      val df = Seq((TestData.injectND(7)(three), TestData.injectND(12)(two))).toDF("three", "two")
-
-      val ct = df.select(
-        rf_convert_cell_type($"three", "uint16ud512") as "three",
-        rf_convert_cell_type($"two", "float32") as "two"
-      )
-
-      val (ct3, ct2) = ct.as[(Tile, Tile)].first()
-
-      ct3.cellType should be (UShortUserDefinedNoDataCellType(512))
-      ct2.cellType should be (FloatConstantNoDataCellType)
-
-      val (cnt3, cnt2) = ct.select(rf_no_data_cells($"three"), rf_no_data_cells($"two")).as[(Long, Long)].first()
-
-      cnt3 should be (7)
-      cnt2 should be (12)
-
-      checkDocs("rf_convert_cell_type")
-    }
-    it("should change NoData value") {
-      val df = Seq((TestData.injectND(7)(three), TestData.injectND(12)(two))).toDF("three", "two")
-
-      val ndCT = df.select(
-        rf_with_no_data($"three", 3) as "three",
-        rf_with_no_data($"two", 2.0) as "two"
-      )
-
-      val (cnt3, cnt2) = ndCT.select(rf_no_data_cells($"three"), rf_no_data_cells($"two")).as[(Long, Long)].first()
-
-      cnt3 should be ((cols * rows) - 7)
-      cnt2 should be ((cols * rows) - 12)
-
-      checkDocs("rf_with_no_data")
-
-      // Should maintain original cell type.
-      ndCT.select(rf_cell_type($"two")).first().withDefaultNoData() should be(ct.withDefaultNoData())
-    }
-  }
 
   describe("arithmetic tile operations") {
     it("should local_add") {
@@ -158,8 +83,11 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
 
       assertEqual(df.selectExpr("rf_local_divide(six, two)").as[ProjectedRasterTile].first(), three)
 
-      assertEqual(df.selectExpr("rf_local_multiply(rf_local_divide(six, 2.0), two)")
-        .as[ProjectedRasterTile].first(), six)
+      assertEqual(
+        df.selectExpr("rf_local_multiply(rf_local_divide(six, 2.0), two)")
+          .as[ProjectedRasterTile]
+          .first(),
+        six)
 
       val maybeThreeTile =
         df.select(rf_local_divide(ExtractTile($"six"), ExtractTile($"two"))).as[Tile]
@@ -221,301 +149,24 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
     }
   }
 
-  describe("tile comparison relations") {
-    it("should evaluate rf_local_less") {
-      val df = Seq((two, three, six)).toDF("two", "three", "six")
-      df.select(rf_tile_sum(rf_local_less($"two", 6))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_less($"two", 1.9))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_less($"two", 2))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_less($"three", $"two"))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_less($"three", $"three"))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_less($"three", $"six"))).first() should be(100.0)
-
-      df.selectExpr("rf_tile_sum(rf_local_less(two, 6))").as[Double].first() should be(100.0)
-      df.selectExpr("rf_tile_sum(rf_local_less(three, three))").as[Double].first() should be(0.0)
-      checkDocs("rf_local_less")
-    }
-
-    it("should evaluate rf_local_less_equal") {
-      val df = Seq((two, three, six)).toDF("two", "three", "six")
-      df.select(rf_tile_sum(rf_local_less_equal($"two", 6))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_less_equal($"two", 1.9))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_less_equal($"two", 2))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_less_equal($"three", $"two"))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_less_equal($"three", $"three"))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_less_equal($"three", $"six"))).first() should be(100.0)
-
-      df.selectExpr("rf_tile_sum(rf_local_less_equal(two, 6))").as[Double].first() should be(100.0)
-      df.selectExpr("rf_tile_sum(rf_local_less_equal(three, three))").as[Double].first() should be(100.0)
-      checkDocs("rf_local_less_equal")
-    }
-
-    it("should evaluate rf_local_greater") {
-      val df = Seq((two, three, six)).toDF("two", "three", "six")
-      df.select(rf_tile_sum(rf_local_greater($"two", 6))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_greater($"two", 1.9))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_greater($"two", 2))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_greater($"three", $"two"))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_greater($"three", $"three"))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_greater($"three", $"six"))).first() should be(0.0)
-
-      df.selectExpr("rf_tile_sum(rf_local_greater(two, 1.9))").as[Double].first() should be(100.0)
-      df.selectExpr("rf_tile_sum(rf_local_greater(three, three))").as[Double].first() should be(0.0)
-      checkDocs("rf_local_greater")
-    }
-
-    it("should evaluate rf_local_greater_equal") {
-      val df = Seq((two, three, six)).toDF("two", "three", "six")
-      df.select(rf_tile_sum(rf_local_greater_equal($"two", 6))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_greater_equal($"two", 1.9))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_greater_equal($"two", 2))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_greater_equal($"three", $"two"))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_greater_equal($"three", $"three"))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_greater_equal($"three", $"six"))).first() should be(0.0)
-      df.selectExpr("rf_tile_sum(rf_local_greater_equal(two, 1.9))").as[Double].first() should be(100.0)
-      df.selectExpr("rf_tile_sum(rf_local_greater_equal(three, three))").as[Double].first() should be(100.0)
-      checkDocs("rf_local_greater_equal")
-    }
-
-    it("should evaluate rf_local_equal") {
-      val df = Seq((two, three, three)).toDF("two", "threeA", "threeB")
-      df.select(rf_tile_sum(rf_local_equal($"two", 2))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_equal($"two", 2.1))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_equal($"two", $"threeA"))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_equal($"threeA", $"threeB"))).first() should be(100.0)
-      df.selectExpr("rf_tile_sum(rf_local_equal(two, 1.9))").as[Double].first() should be(0.0)
-      df.selectExpr("rf_tile_sum(rf_local_equal(threeA, threeB))").as[Double].first() should be(100.0)
-      checkDocs("rf_local_equal")
-    }
-
-    it("should evaluate rf_local_unequal") {
-      val df = Seq((two, three, three)).toDF("two", "threeA", "threeB")
-      df.select(rf_tile_sum(rf_local_unequal($"two", 2))).first() should be(0.0)
-      df.select(rf_tile_sum(rf_local_unequal($"two", 2.1))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_unequal($"two", $"threeA"))).first() should be(100.0)
-      df.select(rf_tile_sum(rf_local_unequal($"threeA", $"threeB"))).first() should be(0.0)
-      df.selectExpr("rf_tile_sum(rf_local_unequal(two, 1.9))").as[Double].first() should be(100.0)
-      df.selectExpr("rf_tile_sum(rf_local_unequal(threeA, threeB))").as[Double].first() should be(0.0)
-      checkDocs("rf_local_unequal")
-    }
-  }
-
-  describe("raster metadata") {
-    it("should get the TileDimensions of a Tile") {
-      val t = Seq(randPRT).toDF("tile").select(rf_dimensions($"tile")).first()
-      t should be (TileDimensions(randPRT.dimensions))
-      checkDocs("rf_dimensions")
-    }
-    it("should get the Extent of a ProjectedRasterTile") {
-      val e = Seq(randPRT).toDF("tile").select(rf_extent($"tile")).first()
-      e should be (extent)
-      checkDocs("rf_extent")
-    }
-
-    it("should get the CRS of a ProjectedRasterTile") {
-      val e = Seq(randPRT).toDF("tile").select(rf_crs($"tile")).first()
-      e should be (crs)
-      checkDocs("rf_crs")
-    }
-
-    it("should parse a CRS from string") {
-      val e = Seq(crs.toProj4String).toDF("crs").select(rf_crs($"crs")).first()
-      e should be (crs)
-    }
-
-    it("should get the Geometry of a ProjectedRasterTile") {
-      val g = Seq(randPRT).toDF("tile").select(rf_geometry($"tile")).first()
-      g should be (extent.jtsGeom)
-      checkDocs("rf_geometry")
-    }
-
-    it("should get the CRS of a RasteRef") {
-      val e = Seq(Tuple1(rasterRef)).toDF("ref").select(rf_crs($"ref")).first()
-      e should be (rasterRef.crs)
-    }
-
-    it("should get the Extent of a RasteRef") {
-      val e = Seq(Tuple1(rasterRef)).toDF("ref").select(rf_extent($"ref")).first()
-      e should be (rasterRef.extent)
-    }
-  }
-
-  describe("per-tile stats") {
-    it("should compute data cell counts") {
-      val df = Seq(TestData.injectND(numND)(two)).toDF("two")
-      df.select(rf_data_cells($"two")).first() shouldBe (cols * rows - numND).toLong
-
-      val df2 = randNDTilesWithNull.toDF("tile")
-      df2.select(rf_data_cells($"tile") as "cells")
-        .agg(sum("cells"))
-        .as[Long]
-        .first() should be (expectedRandData)
-
-      checkDocs("rf_data_cells")
-    }
-    it("should compute no-data cell counts") {
-      val df = Seq(TestData.injectND(numND)(two)).toDF("two")
-      df.select(rf_no_data_cells($"two")).first() should be(numND)
-
-      val df2 = randNDTilesWithNull.toDF("tile")
-      df2.select(rf_no_data_cells($"tile") as "cells")
-        .agg(sum("cells"))
-        .as[Long]
-        .first() should be (expectedRandNoData)
-
-      checkDocs("rf_no_data_cells")
-    }
-
-    it("should properly count data and nodata cells on constant tiles") {
-      val rf = Seq(randPRT).toDF("tile")
-
-      val df = rf
-        .withColumn("make", rf_make_constant_tile(99, 3, 4, ByteConstantNoDataCellType))
-        .withColumn("make2", rf_with_no_data($"make", 99))
-
-      val counts = df.select(
-        rf_no_data_cells($"make").alias("nodata1"),
-        rf_data_cells($"make").alias("data1"),
-        rf_no_data_cells($"make2").alias("nodata2"),
-        rf_data_cells($"make2").alias("data2")
-      ).as[(Long, Long, Long, Long)].first()
-
-      counts should be ((0l, 12l, 12l, 0l))
-    }
-
-    it("should detect no-data tiles") {
-      val df = Seq(nd).toDF("nd")
-      df.select(rf_is_no_data_tile($"nd")).first() should be(true)
-      val df2 = Seq(two).toDF("not_nd")
-      df2.select(rf_is_no_data_tile($"not_nd")).first() should be(false)
-      checkDocs("rf_is_no_data_tile")
-    }
-
-    it("should evaluate exists and for_all") {
-      val df0 = Seq(zero).toDF("tile")
-      df0.select(rf_exists($"tile")).first() should be(false)
-      df0.select(rf_for_all($"tile")).first() should be(false)
-
-      Seq(one).toDF("tile").select(rf_exists($"tile")).first() should be(true)
-      Seq(one).toDF("tile").select(rf_for_all($"tile")).first() should be(true)
-
-      val dfNd = Seq(TestData.injectND(1)(one)).toDF("tile")
-      dfNd.select(rf_exists($"tile")).first() should be(true)
-      dfNd.select(rf_for_all($"tile")).first() should be(false)
-
-      checkDocs("rf_exists")
-      checkDocs("rf_for_all")
-    }
-    it("should find the minimum cell value") {
-      val min = randNDPRT.toArray().filter(c => raster.isData(c)).min.toDouble
-      val df = Seq(randNDPRT).toDF("rand")
-      df.select(rf_tile_min($"rand")).first() should be(min)
-      df.selectExpr("rf_tile_min(rand)").as[Double].first() should be(min)
-      checkDocs("rf_tile_min")
-    }
-
-    it("should find the maximum cell value") {
-      val max = randNDPRT.toArray().filter(c => raster.isData(c)).max.toDouble
-      val df = Seq(randNDPRT).toDF("rand")
-      df.select(rf_tile_max($"rand")).first() should be(max)
-      df.selectExpr("rf_tile_max(rand)").as[Double].first() should be(max)
-      checkDocs("rf_tile_max")
-    }
-    it("should compute the tile mean cell value") {
-      val values = randNDPRT.toArray().filter(c => raster.isData(c))
-      val mean = values.sum.toDouble / values.length
-      val df = Seq(randNDPRT).toDF("rand")
-      df.select(rf_tile_mean($"rand")).first() should be(mean)
-      df.selectExpr("rf_tile_mean(rand)").as[Double].first() should be(mean)
-      checkDocs("rf_tile_mean")
-    }
-
-    it("should compute the tile summary statistics") {
-      val values = randNDPRT.toArray().filter(c => raster.isData(c))
-      val mean = values.sum.toDouble / values.length
-      val df = Seq(randNDPRT).toDF("rand")
-      val stats = df.select(rf_tile_stats($"rand")).first()
-      stats.mean should be (mean +- 0.00001)
-
-      val stats2 = df.selectExpr("rf_tile_stats(rand) as stats")
-        .select($"stats".as[CellStatistics])
-        .first()
-      stats2 should be (stats)
-
-      df.select(rf_tile_stats($"rand") as "stats")
-        .select($"stats.mean").as[Double]
-        .first() should be(mean +- 0.00001)
-      df.selectExpr("rf_tile_stats(rand) as stats")
-        .select($"stats.no_data_cells").as[Long]
-        .first() should be <= (cols * rows - numND).toLong
-
-      val df2 = randNDTilesWithNull.toDF("tile")
-      df2
-        .select(rf_tile_stats($"tile")("data_cells") as "cells")
-        .agg(sum("cells"))
-        .as[Long]
-        .first() should be (expectedRandData)
-
-      checkDocs("rf_tile_stats")
-    }
-
-    it("should compute the tile histogram") {
-      val df = Seq(randNDPRT).toDF("rand")
-      val h1 = df.select(rf_tile_histogram($"rand")).first()
-
-      val h2 = df.selectExpr("rf_tile_histogram(rand) as hist")
-        .select($"hist".as[CellHistogram])
-        .first()
-
-      h1 should be (h2)
-
-      checkDocs("rf_tile_histogram")
-    }
-  }
-
-  describe("array operations") {
-    it("should convert tile into array") {
-      val query = sql(
-        """select rf_tile_to_array_int(
-          |  rf_make_constant_tile(1, 10, 10, 'int8raw')
-          |) as intArray
-          |""".stripMargin)
-      query.as[Array[Int]].first.sum should be (100)
-
-      val tile = FloatConstantTile(1.1f, 10, 10, FloatCellType)
-      val df = Seq[Tile](tile).toDF("tile")
-      val arrayDF = df.select(rf_tile_to_array_double($"tile").as[Array[Double]])
-      arrayDF.first().sum should be (110.0 +- 0.0001)
-
-      checkDocs("rf_tile_to_array_int")
-      checkDocs("rf_tile_to_array_double")
-    }
-
-    it("should convert an array into a tile") {
-      val tile = TestData.randomTile(10, 10, FloatCellType)
-      val df = Seq[Tile](tile, null).toDF("tile")
-      val arrayDF = df.withColumn("tileArray", rf_tile_to_array_double($"tile"))
-
-      val back = arrayDF.withColumn("backToTile", rf_array_to_tile($"tileArray", 10, 10))
-
-      val result = back.select($"backToTile".as[Tile]).first
-
-      assert(result.toArrayDouble() === tile.toArrayDouble())
-
-      // Same round trip, but with SQL expression for rf_array_to_tile
-      val resultSql = arrayDF.selectExpr("rf_array_to_tile(tileArray, 10, 10) as backToTile").as[Tile].first
-
-      assert(resultSql.toArrayDouble() === tile.toArrayDouble())
-
-      val hasNoData = back.withColumn("withNoData", rf_with_no_data($"backToTile", 0))
-
-      val result2 = hasNoData.select($"withNoData".as[Tile]).first
-
-      assert(result2.cellType.asInstanceOf[UserDefinedNoData[_]].noDataValue === 0)
-    }
-  }
-
   describe("analytical transformations") {
+
+    it("should return local data and nodata") {
+      checkDocs("rf_local_data")
+      checkDocs("rf_local_no_data")
+
+      val df = Seq(randNDPRT)
+        .toDF("t")
+        .withColumn("ld", rf_local_data($"t"))
+        .withColumn("lnd", rf_local_no_data($"t"))
+
+      val ndResult = df.select($"lnd").as[Tile].first()
+      ndResult should be(randNDPRT.localUndefined())
+
+      val dResult = df.select($"ld").as[Tile].first()
+      dResult should be(randNDPRT.localDefined())
+    }
+
     it("should compute rf_normalized_difference") {
       val df = Seq((three, two)).toDF("three", "two")
 
@@ -534,14 +185,9 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
     it("should mask one tile against another") {
       val df = Seq[Tile](randPRT).toDF("tile")
 
-      val withMask = df.withColumn("mask",
-        rf_convert_cell_type(
-          rf_local_greater($"tile", 50),
-          "uint8")
-      )
+      val withMask = df.withColumn("mask", rf_convert_cell_type(rf_local_greater($"tile", 50), "uint8"))
 
-      val withMasked = withMask.withColumn("masked",
-        rf_mask($"tile", $"mask"))
+      val withMasked = withMask.withColumn("masked", rf_mask($"tile", $"mask"))
 
       val result = withMasked.agg(rf_agg_no_data_cells($"tile") < rf_agg_no_data_cells($"masked")).as[Boolean]
 
@@ -553,12 +199,11 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
     it("should mask with expected results") {
       val df = Seq((byteArrayTile, maskingTile)).toDF("tile", "mask")
 
-      val withMasked = df.withColumn("masked",
-        rf_mask($"tile", $"mask"))
+      val withMasked = df.withColumn("masked", rf_mask($"tile", $"mask"))
 
       val result: Tile = withMasked.select($"masked").as[Tile].first()
 
-      result.localUndefined().toArray() should be (maskingTile.localUndefined().toArray())
+      result.localUndefined().toArray() should be(maskingTile.localUndefined().toArray())
     }
 
     it("should mask without mutating cell type") {
@@ -568,7 +213,7 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
         .select(rf_cell_type($"masked_tile"))
         .first()
 
-      result should be (byteArrayTile.cellType)
+      result should be(byteArrayTile.cellType)
     }
 
     it("should inverse mask one tile against another") {
@@ -576,12 +221,12 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
 
       val baseND = df.select(rf_agg_no_data_cells($"tile")).first()
 
-      val withMask = df.withColumn("mask",
+      val withMask = df.withColumn(
+        "mask",
         rf_convert_cell_type(
           rf_local_greater($"tile", 50),
           "uint8"
-        )
-      )
+        ))
 
       val withMasked = withMask
         .withColumn("masked", rf_mask($"tile", $"mask"))
@@ -598,16 +243,10 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       val df = Seq[Tile](randPRT).toDF("tile")
       val mask_value = 4
 
-      val withMask = df.withColumn("mask",
-        rf_local_multiply(rf_convert_cell_type(
-          rf_local_greater($"tile", 50),
-          "uint8"),
-          lit(mask_value)
-        )
-      )
+      val withMask =
+        df.withColumn("mask", rf_local_multiply(rf_convert_cell_type(rf_local_greater($"tile", 50), "uint8"), lit(mask_value)))
 
-      val withMasked = withMask.withColumn("masked",
-        rf_mask_by_value($"tile", $"mask", lit(mask_value)))
+      val withMasked = withMask.withColumn("masked", rf_mask_by_value($"tile", $"mask", lit(mask_value)))
 
       val result = withMasked.agg(rf_agg_no_data_cells($"tile") < rf_agg_no_data_cells($"masked")).as[Boolean]
 
@@ -620,34 +259,29 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       val df = Seq((byteArrayTile, maskingTile)).toDF("data", "mask")
 
       // data tile is all data cells
-      df.select(rf_data_cells($"data")).first() should be (byteArrayTile.size)
+      df.select(rf_data_cells($"data")).first() should be(byteArrayTile.size)
 
       // mask by value against 15 should set 3 cell locations to Nodata
       df.withColumn("mbv", rf_mask_by_value($"data", $"mask", 15))
         .select(rf_data_cells($"mbv"))
-        .first() should be (byteArrayTile.size - 3)
+        .first() should be(byteArrayTile.size - 3)
 
       // breaks with issue https://github.com/locationtech/rasterframes/issues/416
-      val result = df.withColumn("mbv", rf_mask_by_value($"data", $"mask", 0))
+      val result = df
+        .withColumn("mbv", rf_mask_by_value($"data", $"mask", 0))
         .select(rf_data_cells($"mbv"))
         .first()
-      result should be (byteArrayTile.size)
+      result should be(byteArrayTile.size)
     }
 
     it("should inverse mask tile by another identified by specified value") {
       val df = Seq[Tile](randPRT).toDF("tile")
       val mask_value = 4
 
-      val withMask = df.withColumn("mask",
-        rf_local_multiply(rf_convert_cell_type(
-          rf_local_greater($"tile", 50),
-          "uint8"),
-          mask_value
-        )
-      )
+      val withMask = df.withColumn("mask", rf_local_multiply(rf_convert_cell_type(rf_local_greater($"tile", 50), "uint8"), mask_value))
 
-      val withMasked = withMask.withColumn("masked",
-        rf_inverse_mask_by_value($"tile", $"mask", mask_value))
+      val withMasked = withMask
+        .withColumn("masked", rf_inverse_mask_by_value($"tile", $"mask", mask_value))
         .withColumn("masked2", rf_mask_by_value($"tile", $"mask", lit(mask_value), true))
 
       val result = withMasked.agg(rf_agg_no_data_cells($"tile") < rf_agg_no_data_cells($"masked")).as[Boolean]
@@ -667,12 +301,12 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
 
       val mask_values = Seq(4, 5, 6, 12)
 
-      val withMasked = df.withColumn("masked",
-        rf_mask_by_values($"tile", $"mask", mask_values:_*))
+      val withMasked = df.withColumn("masked", rf_mask_by_values($"tile", $"mask", mask_values: _*))
 
-      val expected = squareIncrementingPRT.toArray().count(v ⇒ mask_values.contains(v))
+      val expected = squareIncrementingPRT.toArray().count(v => mask_values.contains(v))
 
-      val result = withMasked.agg(rf_agg_no_data_cells($"masked") as "masked_nd")
+      val result = withMasked
+        .agg(rf_agg_no_data_cells($"masked") as "masked_nd")
         .first()
 
       result.getAs[BigInt](0) should be(expected)
@@ -728,7 +362,7 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       checkDocs("rf_abs")
     }
 
-    it("should take logarithms positive cell values"){
+    it("should take logarithms positive cell values") {
       // rf_log10 1000 == 3
       val thousand = TestData.projectedRasterTile(cols, rows, 1000, extent, crs, ShortConstantNoDataCellType)
       val threesDouble = TestData.projectedRasterTile(cols, rows, 3.0, extent, crs, DoubleConstantNoDataCellType)
@@ -740,12 +374,14 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       // ln random tile == rf_log10 random tile / rf_log10(e); random tile square to ensure all positive cell values
       val df2 = Seq(randPositiveDoubleTile).toDF("tile")
       val log10e = math.log10(math.E)
-      assertEqual(df2.select(rf_log($"tile")).as[ProjectedRasterTile].first(),
-                  df2.select(rf_log10($"tile")).as[ProjectedRasterTile].first() / log10e)
+      assertEqual(
+        df2.select(rf_log($"tile")).as[ProjectedRasterTile].first(),
+        df2.select(rf_log10($"tile")).as[ProjectedRasterTile].first() / log10e)
 
       lazy val maybeZeros = df2
         .selectExpr(s"rf_local_subtract(rf_log(tile), rf_local_divide(rf_log10(tile), ${log10e}))")
-        .as[ProjectedRasterTile].first()
+        .as[ProjectedRasterTile]
+        .first()
       assertEqual(maybeZeros, zerosDouble)
 
       // rf_log1p for zeros should be ln(1)
@@ -762,7 +398,7 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
 
     it("should take logarithms with non-positive cell values") {
       val ni_float = TestData.projectedRasterTile(cols, rows, Double.NegativeInfinity, extent, crs, DoubleConstantNoDataCellType)
-      val zero_float =TestData.projectedRasterTile(cols, rows, 0.0, extent, crs, DoubleConstantNoDataCellType)
+      val zero_float = TestData.projectedRasterTile(cols, rows, 0.0, extent, crs, DoubleConstantNoDataCellType)
 
       // tile zeros ==> -Infinity
       val df_0 = Seq(zero).toDF("tile")
@@ -790,39 +426,25 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       )
 
       // base 2
-      assertEqual(
-        df.select(rf_exp2(rf_log2($"tile"))).as[ProjectedRasterTile].first(),
-        six)
+      assertEqual(df.select(rf_exp2(rf_log2($"tile"))).as[ProjectedRasterTile].first(), six)
 
       // base 10
-      assertEqual(
-        df.select(rf_exp10(rf_log10($"tile"))).as[ProjectedRasterTile].first(),
-        six)
+      assertEqual(df.select(rf_exp10(rf_log10($"tile"))).as[ProjectedRasterTile].first(), six)
 
       // plus/minus 1
-      assertEqual(
-        df.select(rf_expm1(rf_log1p($"tile"))).as[ProjectedRasterTile].first(),
-        six)
+      assertEqual(df.select(rf_expm1(rf_log1p($"tile"))).as[ProjectedRasterTile].first(), six)
 
       // SQL
-      assertEqual(
-        df.selectExpr("rf_exp(rf_log(tile))").as[ProjectedRasterTile].first(),
-        six)
+      assertEqual(df.selectExpr("rf_exp(rf_log(tile))").as[ProjectedRasterTile].first(), six)
 
       // SQL base 10
-      assertEqual(
-        df.selectExpr("rf_exp10(rf_log10(tile))").as[ProjectedRasterTile].first(),
-        six)
+      assertEqual(df.selectExpr("rf_exp10(rf_log10(tile))").as[ProjectedRasterTile].first(), six)
 
       // SQL base 2
-      assertEqual(
-        df.selectExpr("rf_exp2(rf_log2(tile))").as[ProjectedRasterTile].first(),
-        six)
+      assertEqual(df.selectExpr("rf_exp2(rf_log2(tile))").as[ProjectedRasterTile].first(), six)
 
       // SQL rf_expm1
-      assertEqual(
-        df.selectExpr("rf_expm1(rf_log1p(tile))").as[ProjectedRasterTile].first(),
-        six)
+      assertEqual(df.selectExpr("rf_expm1(rf_log1p(tile))").as[ProjectedRasterTile].first(), six)
 
       checkDocs("rf_exp")
       checkDocs("rf_exp10")
@@ -834,16 +456,18 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
 
   it("should resample") {
     def lowRes = {
-      def base = ArrayTile(Array(1,2,3,4), 2, 2)
+      def base = ArrayTile(Array(1, 2, 3, 4), 2, 2)
       ProjectedRasterTile(base.convert(ct), extent, crs)
     }
     def upsampled = {
+      // format: off
       def base = ArrayTile(Array(
         1,1,2,2,
         1,1,2,2,
         3,3,4,4,
         3,3,4,4
       ), 4, 4)
+      // format: on
       ProjectedRasterTile(base.convert(ct), extent, crs)
     }
     // a 4, 4 tile to upsample by shape
@@ -866,95 +490,6 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
     checkDocs("rf_resample")
   }
 
-  describe("create encoded representation of 3 band images") {
-
-    it("should create RGB composite") {
-                                        val red = TestData.l8Sample(4).toProjectedRasterTile
-                                        val green = TestData.l8Sample(3).toProjectedRasterTile
-                                        val blue = TestData.l8Sample(2).toProjectedRasterTile
-
-                                        val expected = ArrayMultibandTile(
-                                        red.rescale(0, 255),
-                                        green.rescale(0, 255),
-                                        blue.rescale(0, 255)
-                                        ).color()
-
-                                        val df = Seq((red, green, blue)).toDF("red", "green", "blue")
-
-                                        val expr = df.select(rf_rgb_composite($"red", $"green", $"blue")).as[ProjectedRasterTile]
-
-                                        val nat_color = expr.first()
-
-                                        checkDocs("rf_rgb_composite")
-                                        assertEqual(nat_color.toArrayTile(), expected)
-                                        }
-
-    it("should create an RGB PNG image") {
-                                           val red = TestData.l8Sample(4).toProjectedRasterTile
-                                           val green = TestData.l8Sample(3).toProjectedRasterTile
-                                           val blue = TestData.l8Sample(2).toProjectedRasterTile
-
-                                           val df = Seq((red, green, blue)).toDF("red", "green", "blue")
-
-                                           val expr = df.select(rf_render_png($"red", $"green", $"blue"))
-
-                                           val pngData = expr.first()
-
-                                           val image = ImageIO.read(new ByteArrayInputStream(pngData))
-                                           image.getWidth should be(red.cols)
-                                           image.getHeight should be(red.rows)
-                                           }
-
-    it("should create a color-ramp PNG image") {
-                                                 val red = TestData.l8Sample(4).toProjectedRasterTile
-
-                                                 val df = Seq(red).toDF("red")
-
-                                                 val expr = df.select(rf_render_png($"red", ColorRamps.HeatmapBlueToYellowToRedSpectrum))
-
-                                                 val pngData = expr.first()
-
-                                                 val image = ImageIO.read(new ByteArrayInputStream(pngData))
-                                                 image.getWidth should be(red.cols)
-                                                 image.getHeight should be(red.rows)
-                                                 }
-    it("should interpret cell values with a specified cell type") {
-      checkDocs("rf_interpret_cell_type_as")
-      val df = Seq(randNDPRT).toDF("t")
-        .withColumn("tile", rf_interpret_cell_type_as($"t", "int8raw"))
-      val resultTile = df.select("tile").as[Tile].first()
-
-      resultTile.cellType should be (CellType.fromName("int8raw"))
-      // should have same number of values that are -2 the old ND
-      val countOldNd = df.select(
-        rf_tile_sum(rf_local_equal($"tile", ct.noDataValue)),
-        rf_no_data_cells($"t")
-      ).first()
-      countOldNd._1 should be (countOldNd._2)
-
-      // should not have no data any more (raw type)
-      val countNewNd = df.select(rf_no_data_cells($"tile")).first()
-      countNewNd should be (0L)
-
-    }
-  }
-
-  it("should return local data and nodata"){
-    checkDocs("rf_local_data")
-    checkDocs("rf_local_no_data")
-
-    val df = Seq(randNDPRT).toDF("t")
-      .withColumn("ld", rf_local_data($"t"))
-      .withColumn("lnd", rf_local_no_data($"t"))
-
-    val ndResult = df.select($"lnd").as[Tile].first()
-    ndResult should be (randNDPRT.localUndefined())
-
-    val dResult = df.select($"ld").as[Tile].first()
-    dResult should be (randNDPRT.localDefined())
-  }
-
-
   describe("masking by specific bit values") {
     // Define a dataframe set up similar to the Landsat8 masking scheme
     // Sample of https://www.usgs.gov/media/images/landsat-8-quality-assessment-band-pixel-value-interpretations
@@ -964,60 +499,78 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
     val med_cloud = 2756 // with 1-2 bands saturated
     val hi_cirrus = 6900 // yes cloud, hi conf cloud and hi conf cirrus and 1-2band sat
     val dataColumnCellType = UShortConstantNoDataCellType
-    val tiles = Seq(fill, clear, cirrus, med_cloud, hi_cirrus).map{v ⇒
+    val tiles = Seq(fill, clear, cirrus, med_cloud, hi_cirrus).map { v =>
       (
         TestData.projectedRasterTile(3, 3, 6, TestData.extent, TestData.crs, dataColumnCellType),
-        TestData.projectedRasterTile(3, 3, v, TestData.extent, TestData.crs, UShortCellType) // because masking returns the union of cell types
+        TestData
+          .projectedRasterTile(3, 3, v, TestData.extent, TestData.crs, UShortCellType) // because masking returns the union of cell types
       )
     }
 
-    val df = tiles.toDF("data", "mask")
+    val df = tiles
+      .toDF("data", "mask")
       .withColumn("val", rf_tile_min($"mask"))
 
-    it("should give LHS cell type"){
-      val resultMask = df.select(
-        rf_cell_type(
-          rf_mask($"data", $"mask")
+    it("should give LHS cell type") {
+      val resultMask = df
+        .select(
+          rf_cell_type(
+            rf_mask($"data", $"mask")
+          )
         )
-      ).distinct().collect()
-      all (resultMask) should be (dataColumnCellType)
+        .distinct()
+        .collect()
+      all(resultMask) should be(dataColumnCellType)
 
-      val resultMaskVal = df.select(
-        rf_cell_type(
-          rf_mask_by_value($"data", $"mask", 5)
+      val resultMaskVal = df
+        .select(
+          rf_cell_type(
+            rf_mask_by_value($"data", $"mask", 5)
+          )
         )
-      ).distinct().collect()
+        .distinct()
+        .collect()
 
-      all(resultMaskVal) should be (dataColumnCellType)
+      all(resultMaskVal) should be(dataColumnCellType)
 
-      val resultMaskValues = df.select(
-        rf_cell_type(
-          rf_mask_by_values($"data", $"mask", 5, 6, 7 )
+      val resultMaskValues = df
+        .select(
+          rf_cell_type(
+            rf_mask_by_values($"data", $"mask", 5, 6, 7)
+          )
         )
-      ).distinct().collect()
-      all(resultMaskValues) should be (dataColumnCellType)
+        .distinct()
+        .collect()
+      all(resultMaskValues) should be(dataColumnCellType)
 
-      val resultMaskBit = df.select(
-        rf_cell_type(
-          rf_mask_by_bit($"data", $"mask", 5, true)
+      val resultMaskBit = df
+        .select(
+          rf_cell_type(
+            rf_mask_by_bit($"data", $"mask", 5, true)
+          )
         )
-      ).distinct().collect()
-      all(resultMaskBit) should be (dataColumnCellType)
+        .distinct()
+        .collect()
+      all(resultMaskBit) should be(dataColumnCellType)
 
-      val resultMaskValInv = df.select(
-        rf_cell_type(
-          rf_inverse_mask_by_value($"data", $"mask", 5)
+      val resultMaskValInv = df
+        .select(
+          rf_cell_type(
+            rf_inverse_mask_by_value($"data", $"mask", 5)
+          )
         )
-      ).distinct().collect()
-      all(resultMaskValInv) should be (dataColumnCellType)
+        .distinct()
+        .collect()
+      all(resultMaskValInv) should be(dataColumnCellType)
 
     }
 
-    it("should check values isin"){
+    it("should check values isin") {
       checkDocs("rf_local_is_in")
 
       // tile is 3 by 3 with values, 1 to 9
-      val rf = Seq(byteArrayTile).toDF("t")
+      val rf = Seq(byteArrayTile)
+        .toDF("t")
         .withColumn("one", lit(1))
         .withColumn("five", lit(5))
         .withColumn("ten", lit(10))
@@ -1027,19 +580,19 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
         .withColumn("in_expect_0", rf_local_is_in($"t", array($"ten")))
 
       val e2Result = rf.select(rf_tile_sum($"in_expect_2")).as[Double].first()
-      e2Result should be (2.0)
+      e2Result should be(2.0)
 
       val e1Result = rf.select(rf_tile_sum($"in_expect_1")).as[Double].first()
-      e1Result should be (1.0)
+      e1Result should be(1.0)
 
       val e1aResult = rf.select(rf_tile_sum($"in_expect_1a")).as[Double].first()
-      e1aResult should be (1.0)
+      e1aResult should be(1.0)
 
       val e0Result = rf.select($"in_expect_0").as[Tile].first()
       e0Result.toArray() should contain only (0)
 
     }
-    it("should unpack QA bits"){
+    it("should unpack QA bits") {
       checkDocs("rf_local_extract_bits")
 
       val result = df
@@ -1053,11 +606,12 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       def checker(colName: String, valFilter: Int, assertValue: Int): Unit = {
         // print this so we can see what's happening if something  wrong
         println(s"${colName} should be ${assertValue} for qa val ${valFilter}")
-        result.filter($"val" === lit(valFilter))
+        result
+          .filter($"val" === lit(valFilter))
           .select(col(colName))
           .as[ProjectedRasterTile]
           .first()
-          .get(0, 0) should be (assertValue)
+          .get(0, 0) should be(assertValue)
       }
 
       checker("qa_fill", fill, 1)
@@ -1070,7 +624,9 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       // trivial bits selection (numBits=0) and SQL
       df.filter($"val" === lit(fill))
         .selectExpr("rf_local_extract_bits(mask, 0, 0) AS t")
-        .select(rf_exists($"t")).as[Boolean].first() should be (false)
+        .select(rf_exists($"t"))
+        .as[Boolean]
+        .first() should be(false)
 
       checker("qa_fill", clear, 0)
       checker("qa_cloud", clear, 0)
@@ -1094,7 +650,7 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
         def testIt(): Unit = {
           val tile = projectedRasterTile(3, 3, cellValue, TestData.extent, TestData.crs, cellType)
           val extracted = ExtractBits(tile, bitPosition, numBits)
-          all(extracted.toArray()) should be (expectedValue)
+          all(extracted.toArray()) should be(expectedValue)
         }
       }
 
@@ -1104,44 +660,44 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
         TestCase(ByteCellType, 127, 5, 2, 3),
         TestCase(ByteCellType, -128, 6, 2, 2), // 7th bit is sign
         TestCase(UByteCellType, 255, 6, 2, 3),
-        TestCase(UByteCellType, 255, 10, 2, 0),  // shifting beyond range of cell type results in 0
+        TestCase(UByteCellType, 255, 10, 2, 0), // shifting beyond range of cell type results in 0
         TestCase(ShortCellType, 32767, 15, 1, 0),
         TestCase(ShortCellType, 32767, 14, 2, 1),
         TestCase(ShortUserDefinedNoDataCellType(0), -32768, 14, 2, 2),
         TestCase(UShortCellType, 65535, 14, 2, 3),
-        TestCase(UShortCellType, 65535, 18, 2, 0),  // shifting beyond range of cell type results in 0
+        TestCase(UShortCellType, 65535, 18, 2, 0), // shifting beyond range of cell type results in 0
         TestCase(IntCellType, 2147483647, 30, 2, 1),
         TestCase(IntCellType, 2147483647, 29, 2, 3)
       ).foreach(_.testIt)
 
       // floating point types
-      an [AssertionError] should be thrownBy TestCase[Float](FloatCellType, Float.MaxValue, 29, 2, 3).testIt()
+      an[AssertionError] should be thrownBy TestCase[Float](FloatCellType, Float.MaxValue, 29, 2, 3).testIt()
 
     }
-    it("should mask by QA bits"){
+    it("should mask by QA bits") {
       val result = df
         .withColumn("fill_no", rf_mask_by_bit($"data", $"mask", 0, true))
         .withColumn("sat_0", rf_mask_by_bits($"data", $"mask", 2, 2, 1, 2, 3)) // strict no bands
         .withColumn("sat_2", rf_mask_by_bits($"data", $"mask", 2, 2, 2, 3)) // up to 2 bands contain sat
-        .withColumn("sat_4",
-          rf_mask_by_bits($"data", $"mask", lit(2), lit(2), array(lit(3)))) // up to 4 bands contain sat
+        .withColumn("sat_4", rf_mask_by_bits($"data", $"mask", lit(2), lit(2), array(lit(3)))) // up to 4 bands contain sat
         .withColumn("cloud_no", rf_mask_by_bit($"data", $"mask", lit(4), lit(true)))
         .withColumn("cloud_only", rf_mask_by_bit($"data", $"mask", 4, false)) // mask if *not* cloud
         .withColumn("cloud_conf_low", rf_mask_by_bits($"data", $"mask", lit(5), lit(2), array(lit(0), lit(1))))
         .withColumn("cloud_conf_med", rf_mask_by_bits($"data", $"mask", 5, 2, 0, 1, 2))
         .withColumn("cirrus_med", rf_mask_by_bits($"data", $"mask", 11, 2, 3, 2)) // n.b. this is masking out more likely cirrus.
 
-      result.select(rf_cell_type($"fill_no")).first() should be (dataColumnCellType)
+      result.select(rf_cell_type($"fill_no")).first() should be(dataColumnCellType)
 
       def checker(columnName: String, maskValueFilter: Int, resultIsNoData: Boolean = true): Unit = {
-        /**  in this unit test setup, the `val` column is an integer that the entire row's mask is full of
-          * filter for the maskValueFilter
-          * then check the columnName and look at the masked data tile given by `columnName`
-          * assert that the `columnName` tile is / is not all nodata based on `resultIsNoData`
-        * */
 
-        val printOutcome = if (resultIsNoData) "all NoData cells"
-        else "all data cells"
+        /**  in this unit test setup, the `val` column is an integer that the entire row's mask is full of
+         * filter for the maskValueFilter
+         * then check the columnName and look at the masked data tile given by `columnName`
+         * assert that the `columnName` tile is / is not all nodata based on `resultIsNoData`
+         * */
+        val printOutcome =
+          if (resultIsNoData) "all NoData cells"
+          else "all data cells"
 
         println(s"${columnName} should contain ${printOutcome} for qa val ${maskValueFilter}")
         val resultDf = result
@@ -1156,7 +712,7 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
 //        val celltype = resultDf.select(rf_cell_type(col(columnName))).as[CellType].first()
 //        println(s"Cell type for col ${columnName}: ${celltype}")
 
-        resultToCheck should be (resultIsNoData)
+        resultToCheck should be(resultIsNoData)
       }
 
       checker("fill_no", fill, true)
@@ -1172,14 +728,13 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
       checker("cloud_no", cirrus, false)
     }
 
-    it("mask bits should have SQL equivalent"){
+    it("mask bits should have SQL equivalent") {
 
       df.createOrReplaceTempView("df_maskbits")
 
       val maskedCol = "cloud_conf_med"
       // this is the example in the docs
-      val result = spark.sql(
-        s"""
+      val result = spark.sql(s"""
           |SELECT rf_mask_by_values(
           |            data,
           |            rf_local_extract_bits(mask, 5, 2),
@@ -1188,7 +743,7 @@ class RasterFunctionsSpec extends TestEnvironment with RasterMatchers {
           | FROM df_maskbits
           | WHERE val = 2756
           |""".stripMargin)
-      result.select(rf_is_no_data_tile(col(maskedCol))).first() should be (true)
+      result.select(rf_is_no_data_tile(col(maskedCol))).first() should be(true)
 
     }
   }
